+++ amoc_utac/__init__.py
@@ -25,5 +25,5 @@ References:
 from amoc_utac.constants import GAMMA_AMOC, PACKAGE_REGISTRY_18
 from amoc_utac.system import AmocUTAC
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 __all__ = ["AmocUTAC", "GAMMA_AMOC", "PACKAGE_REGISTRY_18"]